BankLogParser-master/requirements.txt
streamlit==1.32.0
pandas==2.1.3
numpy==1.26.0
scikit-learn==1.4.1.post1
streamlit-option-menu==0.3.6
transformers==4.38.2
torch==2.2.1
torchvision==0.17.1
pdf2image==1.16.3
Pillow==10.2.0
sentencepiece==0.2.0
protobuf==4.25.3
shap==0.48.0
matplotlib==3.8.2
